Make Angular2 Beta 0 happy.
The angular2/angular2 module was dropped, so we've had to find the classes we were importing from it and import them
from the correct new package (usually angular2/common or angular2/core). Apparently we don't need to mention zone.js or
reflect-metadata in webpack.config.js anymore, but I'm not convinced.